+++ packages/react/src/composite/ScrollPane/src/components/ScrollPaneItem/scrollPaneItem.types.ts
@@ -12,15 +12,11 @@ export interface ScrollPaneItemProperties {
   className?: string;
 
   /**
-   * Custom inline styles
-   */
-  style?: React.CSSProperties;
-
-  /**
    * Callback fired when the item is clicked
    */
   onClick?: (event: React.MouseEvent<HTMLDivElement>) => void;
 }
 
-export type ScrollPaneItemProps = ScrollPaneItemProperties &
-  Omit<React.HTMLAttributes<HTMLDivElement>, "children">;
+export type ScrollPaneItemProps = ScrollPaneItemProperties & {
+  style?: React.CSSProperties;
+} & Omit<React.HTMLAttributes<HTMLDivElement>, "children">;